API Reference

General

Lingeling.LglPtr — Type

LglPtr

Wrapper for the Lingeling solver. The object is fully managed by the Lingeling C library but the memory should be released manually using lgl_release.

Fields

ptr::Ptr{Cvoid}: The pointer to the Lingeling solver.

Lingeling.lgl_version — Method

lgl_version()::String

Return the version of the Lingeling SAT-Solver.

Returns

version::String: The version of the Lingeling SAT-Solver as a Julia string.

Lingeling.lgl_init — Method

lgl_init()::LglPtr

Initialize (construct) the Lingeling SAT-Solver.

Returns

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

Lingeling.lgl_release — Method

lgl_release(solver::LglPtr)

Destruct the Lingeling SAT-Solver.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

Options

Lingeling.lgl_setopt — Method

lgl_setopt(solver::LglPtr, opt::String, val::Integer)

Set an option value.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

opt::String: The option name.

val::Integer: The new option value.

Throws

InexactError: If the option value can not be converted to a Cint.

Lingeling.lgl_getopt — Method

lgl_getopt(solver::LglPtr, opt::String)::Cint

Get the current option value.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

opt::String: The option name.

Returns

val::Cint: The option value for the given option.

Lingeling.lgl_defopt — Method

lgl_defopt(solver::LglPtr, opt::String)::Cint

Get the default option value.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

opt::String: The option name.

Returns

val::Cint: The default option value for the given option.

Lingeling.lgl_hasopt — Method

lgl_hasopt(solver::LglPtr, opt::String)::Cint

Check whether an option exists.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

opt::String: The option name.

Returns

val::Cint: Whether the option exists. 1 if it exists, 0 otherwise.

Solving

Lingeling.lgl_add — Method

lgl_add(solver::LglPtr, lit::Integer)

Add a literal of the next clause. A 0 literal terminates the clause.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

lit::Cint: The literal to add. A 0 literal terminates the clause. Negative literals are negated (i.e., -2 means "NOT 2")

Throws

InexactError: If the literal can not be converted to a Cint.

Lingeling.lgl_sat — Method

lgl_sat(solver::LglPtr)::Cint

Call the main SAT routine. The return values are as above, e.g. Lingeling.UNSATISFIABLE, Lingeling.SATISFIABLE, or Lingeling.UNKNOWN.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

Returns

val::Cint: The result of the SAT routine. (e.g. Lingeling.UNSATISFIABLE,

Lingeling.SATISFIABLE, or Lingeling.UNKNOWN)

Lingeling.UNKNOWN — Constant

Return code meaning unknown result.

Lingeling.SATISFIABLE — Constant

Return code meaning the formula is satisfiable.

Lingeling.UNSATISFIABLE — Constant

Return code meaning the formula is unsatisfiable.

Lingeling.lgl_deref — Method

lgl_deref(solver::LglPtr, lit::Integer)::Cint

After 'lgl_sat' was called and returned Lingeling.SATISFIABLE, then the satisfying assignment can be obtained by 'dereferencing' literals. The value of the literal is return as 1 for true, -1 for false and 0 for an unknown value.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

lit::Integer: The literal to dereference.

Returns

val::Cint: The value of the literal. 1 for true, -1 for false, 0 for unknown.

Throws

InexactError: If the literal can not be converted to a Cint.

Lingeling.lgl_freeze — Method

lgl_freeze(solver::LglPtr, lit::Integer)

Freeze a literal meaning that it will not be eliminated during lgl_sat so it can be used in future calls to lgl_add.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

lit::Integer: The literal to freeze.

Throws

InexactError: If the literal can not be converted to a Cint.

Lingeling.lgl_frozen — Method

lgl_frozen(solver::LglPtr, lit::Integer)::Cint

Check whether a literal is frozen.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

lit::Integer: The literal to check.

Returns

val::Cint: Whether the literal is frozen. 1 if it is frozen, 0 otherwise.

Throws

InexactError: If the literal can not be converted to a Cint.

Lingeling.lgl_melt — Method

lgl_melt(solver::LglPtr, lit::Integer)

Melt a literal meaning that it may be eliminated during lgl_sat.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

lit::Integer: The literal to melt.

Throws

InexactError: If the literal can not be converted to a Cint.

Lingeling.lgl_meltall — Method

lgl_meltall(solver::LglPtr)

Melt all literals.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

Lingeling.lgl_usable — Method

lgl_usable(solver::LglPtr, lit::Integer)::Cint

Check whether a literal is usable. A literal becomes unusable if it was not frozen during the last call to lgl_sat. Being unusable means that the literal is not allowed to be used in the next call to lgl_add.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

lit::Integer: The literal to check.

Returns

val::Cint: Whether the literal is usable. 1 if it is usable, 0 otherwise.

Throws

InexactError: If the literal can not be converted to a Cint.

Lingeling.lgl_reusable — Method

lgl_reusable(solver::LglPtr, lit::Integer)::Cint

Check whether a literal is reusable. A literal may not have been frozen during the last call to lgl_sat but could still be reusable (and thus be made usable again using lgl_reuse).

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

lit::Integer: The literal to check.

Returns

val::Cint: Whether the literal is reusable. 1 if it is reusable, 0 otherwise.

Throws

InexactError: If the literal can not be converted to a Cint.

Lingeling.lgl_reuse — Method

lgl_reuse(solver::LglPtr, lit::Integer)

Make a literal usable again. If a literal was not frozen during the last call to lgl_sat but is reusable, it can be made usable again using this function.

Arguments

solver::LglPtr: A pointer to the Lingeling SAT-Solver.

lit::Integer: The literal to make reusable again.

Throws

InexactError: If the literal can not be converted to a Cint.